requirements_hybrid.txt
# Hybrid Detector Requirements
# Install with: pip install -r requirements_hybrid.txt

# Core ML Libraries (existing)
scikit-learn>=1.0.0
pandas>=1.3.0
numpy>=1.21.0

# API & Environment
python-dotenv>=0.19.0
requests>=2.27.0

# Gemini API
google-generativeai>=0.3.0

# Optional: For enhanced features
aiohttp>=3.8.0        # Async HTTP for parallel requests
beautifulsoup4>=4.10.0 # HTML parsing if needed
lxml>=4.9.0           # XML parsing

# Development (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
